pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiflow"
version = "1.3.0"
description = "Enterprise AI Automation Framework"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "BestIxCom Kft", email = "dev@bestixcom.hu"}]
readme = "README.md"

dependencies = [
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "httpx>=0.27",

    # LLM
    "litellm>=1.60",
    "instructor>=1.4",
    "openai>=1.10",

    # Observability
    "langfuse>=2.40",
    "structlog>=24.1",
    "opentelemetry-api>=1.22",
    "opentelemetry-sdk>=1.22",
    "prometheus-client>=0.20",

    # Database & Queue
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "arq>=0.26",
    "redis[hiredis]>=5.0",

    # Security
    "PyJWT[crypto]>=2.8",
    "bcrypt>=4.1",

    # CLI
    "typer>=0.12",
    "rich>=13.7",

    # Prompts
    "pyyaml>=6.0",
    "jinja2>=3.1",

    # Scheduling
    "apscheduler>=4.0.0a5",

    # Notifications
    "aiosmtplib>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "pytest-xdist>=3.5",
    "testcontainers>=4.0",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "detect-secrets>=1.4",
    "libcst>=1.1",
]
vault = ["hvac>=2.1"]
otel = ["opentelemetry-exporter-otlp>=1.22"]
vectorstore = [
    "pgvector>=0.3",
    "tiktoken>=0.7",
    "pymupdf>=1.24",
]
kafka = ["aiokafka>=0.10"]
rabbitmq = ["aio-pika>=9.3"]
azure = ["azure-servicebus>=7.11"]
aws = ["aiobotocore>=2.9"]
ui = ["reflex>=0.6"]
rpa = ["playwright>=1.40"]
local-models = ["sentence-transformers>=3.0"]

[project.scripts]
aiflow = "aiflow.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/aiflow"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = [
    "E501",    # line-too-long — enforced by ruff format, not check
    "TC001",   # typing-only-first-party-import — large refactor, low priority
    "TC002",   # typing-only-third-party-import
    "TC003",   # typing-only-standard-library-import
    "SIM105",  # suppressible-exception — contextlib.suppress preference is stylistic
    "UP042",   # replace-str-enum — requires StrEnum migration
    "UP046",   # non-pep695-generic-class — Python 3.12+ syntax, opt-in later
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B017"]     # assert-raises-exception OK in tests
"skills/**/*.py" = [
    "N806",  # non-lowercase-variable-in-function (domain naming: TF-IDF, BPMN, etc.)
    "N803",  # invalid-argument-name (domain params)
    "E402",  # module-import-not-at-top (conditional imports in skills)
    "E702",  # multiple-statements-on-one-line (reference/example code)
    "E741",  # ambiguous-variable-name (l, O in ML code)
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring Docker services",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["aiflow"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
